add explicit conf paths and binaries
by doing this we can create scripts that can know exactly what each config. value is.

for example, if we're writing a script that needs to know where the apiserver bin is for ck8s, the script can parse the yaml and find master.bins[0] and know that our binary name for apiserver is kube-apiserver